US4239: Move fields into service
Moved fields/model into service. 
Got rid of redundant getAgeRanges() call in group service.
