+++ crossroads.net/app/group_tool/index.js
@@ -1,6 +1,7 @@
 import CONSTANTS from 'crds-constants';
 import ParticipantService from './services/participant.service';
 import GroupService from './services/group.service';
+import CreateGroupService from './services/createGroup.service';
 import groupToolRouter from './groupTool.routes';
 import groupToolFormlyBuilderConfig from './groupTool.formlyConfig'
 import './formlyWrappers/createGroupWrapper.html'
@@ -11,7 +12,8 @@ export default angular.
   config(groupToolRouter).
   config(groupToolFormlyBuilderConfig).
   service('ParticipantService', ParticipantService).
-  service('GroupService', GroupService)
+  service('GroupService', GroupService).
+  service('CreateGroupService', CreateGroupService)
   ;
 
 import myGroups from './my_groups';
