+++ crossroads.net/app/group_tool/services/index.js
@@ -1,8 +1,10 @@
 import participantService from './participant.service';
 import groupService from './group.service';
+import createGroupService from './createGroup.service';
 import CONSTANTS from 'crds-constants';
 
 export default angular.
   module(CONSTANTS.MODULES.GROUP_TOOL).
   service('ParticipantService', participantService).
-  service('GroupService', groupService);
+  service('GroupService', groupService).
+  service('CreateGroupService', createGroupService);
